cleanup + move sql parsing/modifications into its own file
